% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up_subfunctions.R
\name{age_parameters_human_func}
\alias{age_parameters_human_func}
\title{Calculate key age parameters to for human population}
\usage{
age_parameters_human_func(number_age_classes)
}
\arguments{
\item{number_age_classes}{the number of age classes (humans)}
}
\value{
age rate and age width
}
\description{
Calculate age rate and age widths for human compartments depending on number of age groups
}
