% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up_subfunctions.R
\name{age_parameters_pig_func}
\alias{age_parameters_pig_func}
\title{Calculate key age parameters to inform life tables (for pig population)}
\usage{
age_parameters_pig_func(number_age_classes_pig, pig_age_class_width)
}
\arguments{
\item{number_age_classes_pig}{the number of age classes (pigs)}

\item{pig_age_class_width}{vector for mean no. of months in each age compartment}
}
\value{
age rate and age width
}
\description{
Calculate age rate and age widths for pig compartments depending on number of age groups
}
