% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{age_struc_pig_vacc_func}
\alias{age_struc_pig_vacc_func}
\title{age_struc_pig_vacc_func}
\usage{
age_struc_pig_vacc_func(oldest_age, intervention_frequency)
}
\arguments{
\item{oldest_age}{oldest_age - upper limit of age vector}

\item{Intervention_frequency}{and 2 months to this to define lowest limit for age vector}
}
\value{
age_target_pig_vaccine : vector defining lower and upper age limits for pig vaccination
}
\description{
define age structure for pig vaccine if no age structure included (to account for vaccination from 2 months, and interval between 1st + 2nd dose which must be < 4 months)
}
