% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence_adjustment.R
\name{apparent_prevalence_packaging_func}
\alias{apparent_prevalence_packaging_func}
\title{Apparent prevalence packaging}
\usage{
apparent_prevalence_packaging_func(sens, spec, TP)
}
\arguments{
\item{sens}{sensitivity of diagnostic}

\item{spec}{specificity of diagnostic}

\item{TP}{model generated prevalence}
}
\value{
Apparent / observed prevalence
}
\description{
uses Prev_adjustment_AP_func to generate a column with observed/ apparent prevalence in model run dataframe
}
