% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{beta_equilibrium}
\alias{beta_equilibrium}
\title{Pig to human transmission paramter}
\usage{
beta_equilibrium(
  alpha,
  IH0_all,
  IHC0_all,
  eta,
  dH,
  IP0_all,
  PPS,
  SH0_all,
  SHC0_all
)
}
\arguments{
\item{alpha}{Human recovery rate from Taenisasis (per month)}

\item{eta}{Human recovery rate from Cysticercosis (oer month)}

\item{dH}{Human mortality rate (per month)}

\item{PPS}{Pig popultion size}

\item{IH0}{Initial number of Human: T +  C-}

\item{IHC0}{Initial number of Human: T +  C +}

\item{IP0}{Initial number of infected pigs}

\item{SH0}{Initial number of Human: susceptible}

\item{SHC0}{Initial number of Human: T- C +}
}
\value{
The equilibrium pig to human transmission parameter
}
\description{
Calculate the transmission parameter (pig->human) at equiliberium
}
