% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Check_effect}
\alias{Check_effect}
\title{Check intervention effect input}
\usage{
Check_effect(Intervention_effect)
}
\arguments{
\item{Intevention}{Vector of one or more parameter interventions}
}
\description{
Checks that the intervention effect inputs are correct
}

