% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{check_interventions}
\alias{check_interventions}
\title{Check intervention input}
\usage{
check_interventions(intervention_stage2)

check_interventions(intervention_stage2)

check_interventions(intervention_stage2)
}
\arguments{
\item{Intevention}{Vector of one or more parameter interventions}
}
\description{
Checks that the interventions input are correct

Checks that the interventions input are correct

Checks that the interventions input are correct
}
