% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{check_interventions_stg1}
\alias{check_interventions_stg1}
\title{Check intervention input (stage 1 of multi-stage intervention)}
\usage{
check_interventions_stg1(intervention_stage1)
}
\arguments{
\item{intervention_stage1}{Vector of one or more parameter interventions (stage 1)}
}
\description{
Checks that the interventions input are correct
}
