% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{check_interventions_stg2}
\alias{check_interventions_stg2}
\title{Check intervention input (stage 2 of multi-stage intervention)}
\usage{
check_interventions_stg2(intervention_stage2)
}
\arguments{
\item{intervention_stage2}{Vector of one or more parameter interventions (stage 2)}
}
\description{
Checks that the interventions input are correct
}
