% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_wrappers.R
\name{clean_initial_states}
\alias{clean_initial_states}
\title{Clean initial states}
\usage{
clean_initial_states(x)
}
\arguments{
\item{x}{dataframe of model output states}
}
\value{
Initial states formatted for input (without summary columns)
}
\description{
Clean initial states
}
