% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{e0_equilibrium}
\alias{e0_equilibrium}
\title{Equilibrium Eggs}
\usage{
e0_equilibrium(delta, HPS, TPrev, CPrev, dE)
}
\arguments{
\item{delta}{Egg production rate (per month)}

\item{HPS}{Human population size}

\item{TPrev}{Taeniasis prevalence in human population}

\item{CPrev}{Cysticercosis prevalence in human population}

\item{dE}{The egg mortlality / removal rate (per month)}
}
\value{
The Equilbirum number of eggs in the environment
}
\description{
Calculate the number of eggs in the environment at equlibrium
}
