% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up_subfunctions.R
\name{human_age_class_proportions_func}
\alias{human_age_class_proportions_func}
\title{Calculate proportions of humnas in each age class}
\usage{
human_age_class_proportions_func(
  age_rate,
  na_human,
  dH,
  HPS,
  SHC0_total,
  IH0_total,
  IHC0_total
)
}
\arguments{
\item{age_rate}{age rate for humans between age classes}

\item{na_human}{number of human age classes}

\item{dH}{natural death rate of humans}

\item{HPS}{human populations size}

\item{SHC0_total}{susceptible (taeniasis) human}

\item{IH0_total}{infected humans (taeniasis) at t0}

\item{IHC0_total}{infected humans (cysticercosis) at t0}
}
\value{
proportions of humans in each age class and state
}
\description{
Calculate proprtion of humans in each age class for each state using age rates
}
