% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{intervention_effect_size}
\alias{intervention_effect_size}
\title{Pre-set intervention effects}
\usage{
intervention_effect_size()
}
\value{
A list Intervention effects
}
\description{
Provides preset intervention effect
}
