% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{intervention_effect_size_set_up}
\alias{intervention_effect_size_set_up}
\title{Coverage user defined intervention effect size}
\usage{
intervention_effect_size_set_up(
  pig_MDA_cov,
  pig_vaccine_ds1_cov,
  pig_vaccine_ds2_cov,
  human_testtreat_cov,
  human_MDAnic_cov,
  human_MDApzq_cov,
  pig_ofz_efficacy,
  human_pzq_efficacy,
  human_nic_efficacy,
  pig_MDA_prop_noimmunity
)
}
\arguments{
\item{pig_MDA_cov}{pig treatment coverage}

\item{pig_vaccine_ds1_cov}{pig vaccine coverage (dose 1)}

\item{pig_vaccine_ds2_cov}{pig vaccine coverage (dose 2)}

\item{human_testtreat_cov}{human test and treat coverage}

\item{human_MDAnic_cov}{human MDA (with niclosamide) coverage}

\item{human_MDApzq_cov}{human MDA (with praziquantel) coverage}

\item{pig_ofz_efficacy}{efficacy for oxfendazole for pig MDA}

\item{human_pzq_efficacy}{efficacy for praziqunatel for human MDA}

\item{human_nic_efficacy}{efficacy for niclosamide for human MDA}

\item{pig_MDA_prop_noimmunity}{proportion of pigs without immunity following treatment}
}
\value{
A list Intervention effects
}
\description{
enables user specified coverage to create list of proportions for state intervention moves
}
