% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{intervention_event_param}
\alias{intervention_event_param}
\title{Implement parameter intervention}
\usage{
intervention_event_param(params, intervention, intervention_effect)
}
\arguments{
\item{params}{List of model parameters}

\item{intervention}{Vector of one or more parameter interventions}

\item{intervention_effect}{a list of intervention effect sizes}
}
\description{
Implements one or more interventions that involve a parameter value being altered
}
