% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Intervention_event_state}
\alias{Intervention_event_state}
\title{Implement parameter intervention}
\usage{
Intervention_event_state(States, Intervention, Intervention_effect)
}
\arguments{
\item{States}{List of model states}

\item{Intervention_effect}{a list of intervention effect sizes}

\item{Intevention}{Vector of one or more parameter interventions}
}
\description{
Implements one or more interventions that involve a parameter value being altered
}
