% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up_subfunctions.R
\name{life_tables_pigs_func}
\alias{life_tables_pigs_func}
\title{Calculate life table for pig population}
\usage{
life_tables_pigs_func(
  number_age_classes_pig,
  slaughter_age_min,
  slgtage,
  slgtage_bfr,
  dP = dP,
  dPslg = dPslg,
  na_pig
)
}
\arguments{
\item{number_age_classes_pig}{the number of age classes (pigs)}

\item{slaughter_age_min}{the age from which pigs are slaughtered (months), user-specified}

\item{slgtage}{age-class where slaughter rate first applied to}

\item{slgtage_bfr}{age-class before slaughter age first begins}

\item{dP}{pig death rate (natural)}

\item{dPslg}{pig slaughter-age death rate}

\item{na_pig}{number age classes for pigs (processed)}
}
\value{
life table output
}
\description{
To set up population distribution at baseline: LIFE TABLES APPROACH for Age-structured model 
Based on code from Verity et al: https://github.com/mrc-ide/covfefe 
Further reference here FAO paper: http://www.fao.org/tempref/docrep/fao/008/a0212e/a0212E09.pdf
}
