% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up.R
\name{month_rate}
\alias{month_rate}
\title{Monthly rate}
\usage{
month_rate(dur)
}
\arguments{
\item{dur}{Average duration (years)}
}
\value{
Monthly rate
}
\description{
Converts an average duration (years) into a monthly rate
}

