% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Move_state}
\alias{Move_state}
\title{Move individuals between state variables}
\usage{
Move_state(States, from, to, proportion)
}
\arguments{
\item{States}{List of state variable}

\item{from}{The name of the state variabe individuals will be moved out of}

\item{to}{The name of the state variable that individuals will be moved in to}

\item{proportion}{The propotion of individuals in from that get moved into to}
}
\description{
Moves a propotion of individuals from one state compartemnt to another
}

