% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{move_state_double}
\alias{move_state_double}
\title{Move individuals between state variables two flows}
\usage{
move_state_double(states, from, to_1, to_2, proportion_1, proportion_2)
}
\arguments{
\item{from}{The name of the state variabe individuals will be moved out of}

\item{States}{List of state variable}

\item{to}{The name of the state variable that individuals will be moved in to}

\item{proportion}{The propotion of individuals in from that get moved into to}
}
\description{
Moves a propotions of individuals from one state compartemnt to two others
}
