% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{move_state_triple}
\alias{move_state_triple}
\title{Move individuals between state variables three flows}
\usage{
move_state_triple(
  states,
  from,
  to_1,
  to_2,
  to_3,
  proportion_1,
  proportion_2,
  proportion_3
)
}
\arguments{
\item{states}{List of state variable}

\item{from}{The name of the state variabe individuals will be moved out of}

\item{to_1}{The name of the state variable that individuals will be moved in to (first step)}

\item{to_2}{The name of the state variable that individuals will be moved in to (second step)}

\item{to_3}{The name of the state variable that individuals will be moved in to (third step)}

\item{proportion_1}{The proportion of individuals in from that get moved into to (step 1)}

\item{proportion_2}{The proportion of individuals in from that get moved into to (step 2)}

\item{proportion_3}{The proportion of individuals in from that get moved into to (step 3)}
}
\description{
Moves a propotions of individuals from one state compartemnt to three others
}
