% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up_subfunctions.R
\name{Pig_age_class_proportions_func}
\alias{Pig_age_class_proportions_func}
\title{Calculate proportions of pigs in each age class}
\usage{
Pig_age_class_proportions_func(
  PPS,
  pig_demography,
  na_pig,
  IPL0_total,
  IPH0_total
)
}
\arguments{
\item{PPS}{pig population size}

\item{pig_demography}{pig life-table}

\item{na_pig}{number age classes for pigs (processed)}

\item{IPL0_total}{total number of infected (low burden) pigs}

\item{IPH0_total}{total number of infected (high burden) pigs}
}
\value{
proportions of pigs in each age class and state
}
\description{
Using age stable values from life table, calculate proprtion of pigs in each age class for each state
}
