% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up.R
\name{pil_equilibrium}
\alias{pil_equilibrium}
\title{Low intensity infected pig -> human infection probability}
\usage{
pil_equilibrium(beta, chi, phi)
}
\arguments{
\item{chi}{rate of eating a pork meal (per month)}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}

\item{Beta}{Pork to human tranmission parameter}

\item{pil}{Low intensity infected pig -> human contact rate (per month)}

\item{pih}{High intensity infected pig -> human contact rate (per month)}

\item{Beta}{Pork to human tranmission parameter}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}
}
\value{
The equilibrium transmission probability between a low intesnisty infected pig and human

The equilibrium transmission probability between a low intesnisty infected pig and human
}
\description{
Calculate the transmission probability between a low intensity infected pig and human at equiliberium

Calculate the transmission probability between a low intensity infected pork meal and human at equiliberium
}

