% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{PRCC}
\alias{PRCC}
\title{PRCC}
\usage{
PRCC(outcome, covariates)
}
\arguments{
\item{outcome}{a vector of output (e.g. cumulative cysticercosis cases), with one value for each set of parameter draws}

\item{covariates}{a dataframe of the input parameters (one row for each draw)}
}
\value{
A vector of PRCC scores, one for each inoput covariate
}
\description{
Calculates the PRCC statistic
}

