% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{PRCC_sig}
\alias{PRCC_sig}
\title{PRCC significance test}
\usage{
PRCC_sig(prcc_values, N, N_par)
}
\arguments{
\item{prcc_values}{a vector of prcc scores}

\item{N}{The number of draws}

\item{N_par}{The number of covariates}
}
\value{
A vector of PRCC scores, one for each inoput covariate
}
\description{
Calculates the test statistic for PRCC scores
}

