% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{pre_pig_vaccine}
\alias{pre_pig_vaccine}
\title{pre_pig_vaccine}
\usage{
pre_pig_vaccine(age_target, tail_states)
}
\arguments{
\item{age_target}{numeric of specific age groups to target/ select}

\item{tail_states}{output from Inter_run_processing}
}
\value{
tail states targetted for age-structured pig vaccine intervention
}
\description{
Takes processed tail states (following burn-in and a run) and selects specific age groups to implement pig vaccine moves
}
