% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence_adjustment.R
\name{prev_adjustment_AP_func}
\alias{prev_adjustment_AP_func}
\title{Apparent prevalence adjustment}
\usage{
prev_adjustment_AP_func(sens, spec, TP)
}
\arguments{
\item{sens}{sensitivty of diagnostic}

\item{spec}{specifcity of diagnostic}

\item{TP}{model generated prevalence}
}
\value{
Apparent / observed prevalence
}
\description{
estimates apparent prevalence from true underlying model predicted prevalence
}
