% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence_adjustment.R
\name{prev_adjustment_TP_func}
\alias{prev_adjustment_TP_func}
\title{Adjustment observed prevalence to true prevalence}
\usage{
prev_adjustment_TP_func(sens, spec, positive, total)
}
\arguments{
\item{sens}{sensitivty of diagnostic}

\item{spec}{specifcity of diagnostic}

\item{positive}{number of positive observed}

\item{total}{total sampled}
}
\value{
True prevalence median from MCMC
}
\description{
Uses MCMC from prevalence package to adjust observed positives/ total sampled and diagnostic se and sp to true prevalence
see for more info: http://prevalence.cbra.be/?main=download
}
