% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0.R
\name{R0}
\alias{R0}
\title{R0 wrapper}
\usage{
R0(Params)
}
\arguments{
\item{Params}{A list of model parameters: see \code{\link{R0_internal}} for list}
}
\description{
A wrapper function to faciloitate the R0 function being called with a list of parameters
}

