% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{replace_param}
\alias{replace_param}
\title{Change parameter}
\usage{
replace_param(params, param_name, effect_size)
}
\arguments{
\item{params}{List of model parameters}

\item{param_name}{The name of the parameter to be changed}

\item{effect_size}{The intervention effect size. Where the new parameter value = old value multiplied by the effect size.}
}
\description{
Replaces a paramter with an aletred values after a parameter event
}
