% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{Sensitivity_intervention}
\alias{Sensitivity_intervention}
\title{Sensitivity draws for interventions}
\usage{
Sensitivity_intervention(N)
}
\arguments{
\item{N}{The number of draws to produce}
}
\value{
A list of intervention effect size sets
}
\description{
Simulates interventions effect sensivities
}

