% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{Sensitivity_params}
\alias{Sensitivity_params}
\title{Sensitivity parameter LHS draw}
\usage{
Sensitivity_params(N)
}
\arguments{
\item{N}{The number of draws to produce}
}
\value{
A list of parameter sets
}
\description{
Simulates parameter sets from a Latin Hypercube Sample (uniform distribution)
}
