% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up.R
\name{set_up_analytic}
\alias{set_up_analytic}
\title{Set up (age-structured model), analytic coefficients}
\usage{
set_up_analytic(
  LEP = 10,
  slgEP = 1,
  HPS = 10000,
  PPS = 2000,
  AEL = 2,
  delta = 960000,
  tau_input = NULL,
  theta_input = NULL,
  beta_input = NULL,
  psi = 0.5,
  ATL = 2,
  ADI = 3,
  LEH = 54,
  phi = 0.8,
  chi = 0.5,
  RR_cysticercosis = 1,
  epsilon = 0.01,
  RR_infection = 1,
  RR_consumption = -0.25,
  number_age_classes_pig = 150,
  slaughter_age_min = 6,
  pig_age_class_width = 1,
  number_age_classes_human = 7,
  PCPrev = 0.2,
  PC_sens = NULL,
  PC_spec = NULL,
  CPrev = 0.07,
  C_sens = NULL,
  C_spec = NULL,
  TPrev = 0.02,
  T_sens = NULL,
  T_spec = NULL
)
}
\arguments{
\item{LEP}{Pig life expectancy (years)}

\item{slgEP}{life expectancy (years) after pigs reach slaughter age}

\item{HPS}{Human population size}

\item{PPS}{Pig population size}

\item{AEL}{Average life expectancy of egg in the environment (weeks)}

\item{delta}{Egg production rate (per month)}

\item{tau_input}{user specified egg to pig transmission coefficient}

\item{theta_input}{user specified egg to human transmission coefficient}

\item{beta_input}{user specified pig to human transmission coefficient}

\item{psi}{Pig rate of transition from prepatent to infectious (average duration for maturation of cysts)}

\item{ATL}{Average life expectancy of T. solium adult worm in human (years)}

\item{ADI}{Average duration of cycticercosis infection in humans (years)}

\item{LEH}{Average life expectancy of a human (years)}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}

\item{chi}{Average number of pork meals (200g portion) per person per month}

\item{RR_cysticercosis}{Risk multiplier for Cysticercosis if human has Taeniasis}

\item{epsilon}{Pig rate of loss of naturally aquired immunity}

\item{RR_infection}{Increased RR of infection given consumption of high cyct burden pork, compared with low cyst burden}

\item{RR_consumption}{Decreased RR of consumption of high cyst burden pork, compared with low cyst burden pork}

\item{number_age_classes_pig}{Number of age classes for pig population structure}

\item{slaughter_age_min}{Minimum age (in months) from which pigs are slaughtered}

\item{pig_age_class_width}{duration/width of each pig age class (months)}

\item{number_age_classes_human}{number of age classes for human population structure}

\item{PCPrev}{Cysticercosis prevalence in pigs (true underlying prevalence)}

\item{PC_sens}{Sensitivity of pig cysticercosis diagnostic}

\item{PC_spec}{Specificity of pig cysticercosis diagnostic}

\item{CPrev}{Cysticercosis prevalence in humans (true underlying prevalence)}

\item{C_sens}{Sensitivity of human cysticercosis diagnostic}

\item{C_spec}{Specificity of human cysticercosis diagnostic}

\item{TPrev}{Taeniasis prevalence (true underlying prevalence)}

\item{T_sens}{Sensitivity of human taeniasis diagnostic}

\item{T_spec}{Specificity of human taeniasis diagnostic}
}
\value{
A list of two lists: model parameters, and initial state values.
}
\description{
Calculates internal parameters and equilibrium starting values for state
variables, deriving the transmission coefficients in closed form from the
requested prevalences.
}
\details{
The coefficients are obtained by inverting a reduced form of the
transmission system. The integrated age-structured model therefore settles
at an equilibrium somewhat below the requested prevalences: with
`PCPrev = 0.25` the porcine prevalence settles near 0.15. Supply
`tau_input`, `beta_input` and `theta_input` to override the derived values,
or use [set_up()] with `tune = TRUE` to have them solved against the
integrated system.
}
\seealso{
[set_up()] for the tuned wrapper, [tune_transmission()] for the
  solver.
}
