% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_wrappers.R
\name{single_run}
\alias{single_run}
\title{run Cysticercosis model ODE}
\usage{
single_run(tt, params, states)
}
\arguments{
\item{tt}{vector of times}

\item{params}{list of parameters}

\item{states}{list of states}
}
\description{
runs a single implementation of the ODE Cysticercosis model
}
