% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{tau_equilibrium}
\alias{tau_equilibrium}
\title{Egg to pig transmission paramter}
\usage{
tau_equilibrium(dP, dPslg, IP0_nonslgt, IP0_slgt, IP0_all, SP0_all, E0)
}
\arguments{
\item{dP}{Pig mortality rate (per month)}

\item{E0}{The equilibrium number of eggs in the environment}

\item{IP0}{Initial number of infected pigs}

\item{SP0}{Intitial number of susceptible pigs}
}
\value{
The equilibrium egg to pig transmission parameter
}
\description{
Calculate the transmission parameter (egg->pig) at equiliberium
}
