% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{theta_equilibrium}
\alias{theta_equilibrium}
\title{Egg to human transmission paramter}
\usage{
theta_equilibrium(bH, eta, SHC0_all, IHC0_all, dH, SH0_all, IH0_all, E0, RR)
}
\arguments{
\item{bH}{Births of humans (net)}

\item{eta}{Human recovery rate from Cysticercosis (oer month)}

\item{dH}{Human mortality rate (per month)}

\item{E0}{The equilibrium number of eggs in the environment}

\item{RR}{Risk multiplier for Cysticercosis if human has Taeniasis}

\item{SHC0}{Initial number of Human: T- C +}

\item{IHC0}{Initial number of Human: T +  C +}

\item{SH0}{Initial number of Human: susceptible}

\item{IH0}{Initial number of Human: T +  C-}
}
\value{
The equilibrium egg to human transmission parameter
}
\description{
Calculate the transmission parameter (egg->human) at equiliberium
}
