% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{update_states}
\alias{update_states}
\title{update_states}
\usage{
update_states(states_move, tail_states)
}
\arguments{
\item{states_move}{new age-specific states (updated following intervention moves)}

\item{tail_states}{output from Inter_run_processing (to update specific age-classes)}
}
\value{
overall tail states list with new / updated specific age - classes following intervention moves
}
\description{
Identifies age targetted states and updates these specific states in the overall tail states from the initial model run
}
