% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{Beta_equilibrium}
\alias{Beta_equilibrium}
\title{Pig to human transmission paramter}
\usage{
Beta_equilibrium(alpha, IH0, IHC0, eta, dH, IP0, PPS, SH0, SHC0)
}
\arguments{
\item{alpha}{Human recovery rate from Taenisasis (per month)}

\item{IH0}{Initial number of Human: T+ C-}

\item{IHC0}{Initial number of Human: T+ C+}

\item{eta}{Human recovery rate from Cysticercosis (oer month)}

\item{dH}{Human mortality rate (per month)}

\item{IP0}{Initial number of infected pigs}

\item{PPS}{Pig popultion size}

\item{SH0}{Initial number of Human: susceptible}

\item{SHC0}{Initial number of Human: T- C+}
}
\value{
The equilibrium pig to human transmission parameter
}
\description{
Calculate the transmission parameter (pig->human) at equiliberium
}

