% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Check_interventions}
\alias{Check_interventions}
\title{Check intervention input}
\usage{
Check_interventions(Intervention)
}
\arguments{
\item{Intervention}{Vector of one or more parameter interventions}
}
\description{
Checks that the interventions input are correct
}
