% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_run_processing.R
\name{Inter_run_setup}
\alias{Inter_run_setup}
\title{Inter_run_setup}
\usage{
Inter_run_setup(model_output, na_pig, na_human)
}
\arguments{
\item{model_output}{large desolve element generated from odin run (ODE output)}

\item{na_pig}{number of age classes}

\item{na_human}{number of age classes (human)}
}
\value{
single lists of state variable values
}
\description{
Converts output from odin into Initial_states (all initial states and converts age-strcutured states & age_rates into vectors) to re-supply odin
}
