% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Intervention_effect_size}
\alias{Intervention_effect_size}
\title{Pre-set intervention effects}
\usage{
Intervention_effect_size()
}
\value{
A list Intervention effects
}
\description{
Provides preset intervention effect
}
