% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{intervention_effect_size_set_up}
\alias{intervention_effect_size_set_up}
\title{Coverage user defined intervention effect size}
\usage{
intervention_effect_size_set_up(
  pig_MDA_cov,
  pig_vaccine_ds1_cov,
  pig_vaccine_ds2_cov,
  human_testtreat_cov,
  human_MDAnic_cov,
  human_MDApzq_cov,
  pig_MDA_prop_noimmunity
)
}
\value{
A list Intervention effects
}
\description{
enables user specified coverage to create list of proportions for state intervention moves
}
