% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Intervention_event_param}
\alias{Intervention_event_param}
\title{Implement parameter intervention}
\usage{
Intervention_event_param(Params, Intervention, Intervention_effect)
}
\arguments{
\item{Params}{List of model parameters}

\item{Intervention}{Vector of one or more parameter interventions}

\item{Intervention_effect}{a list of intervention effect sizes}
}
\description{
Implements one or more interventions that involve a parameter value being altered
}
