% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{intervention_event_state}
\alias{intervention_event_state}
\title{Implement parameter intervention}
\usage{
intervention_event_state(states, intervention, intervention_effect)
}
\arguments{
\item{states}{List of model states}

\item{Intevention}{Vector of one or more parameter interventions}

\item{Intervention_effect}{a list of intervention effect sizes}
}
\description{
Implements one or more interventions that involve a parameter value being altered
}
