% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{move_state}
\alias{move_state}
\title{Move individuals between state variables}
\usage{
move_state(states, from, to, proportion)
}
\arguments{
\item{states}{List of state variable}

\item{from}{The name of the state variabe individuals will be moved out of}

\item{to}{The name of the state variable that individuals will be moved in to}

\item{proportion}{The propotion of individuals in from that get moved into to}
}
\description{
Moves a propotion of individuals from one state compartemnt to another
}
