% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_events.R
\name{Move_state_double}
\alias{Move_state_double}
\title{Move individuals between state variables two flows}
\usage{
Move_state_double(States, from, to_1, to_2, proportion_1, proportion_2)
}
\arguments{
\item{States}{List of state variable}

\item{from}{The name of the state variabe individuals will be moved out of}

\item{to_1}{The name of the state variable that individuals will be moved in to}

\item{to_2}{The name of the state variable that individuals will be moved in to}

\item{proportion_1}{The propotion of individuals in from that get moved into to}

\item{proportion_2}{The propotion of individuals in from that get moved into to}
}
\description{
Moves a propotions of individuals from one state compartemnt to two others
}
