% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium_solutions.R
\name{pil_equilibrium}
\alias{pil_equilibrium}
\title{Low intensity infected pig -> human infection probability}
\usage{
pil_equilibrium(beta, chi, phi, CRR)
}
\arguments{
\item{chi}{rate of aquiring a pork meal (per month)}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}

\item{CRR}{Combined reltive risk for high cyst burden pork (increased RR of infection * decreased RR of consumption)}

\item{Beta}{Pork to human tranmission parameter}
}
\value{
The equilibrium transmission probability between a low intesnisty infected pig and human
}
\description{
Calculate the transmission probability between a low intensity infected pork meal and human at equiliberium
}

