% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{pre_human_test_and_treat}
\alias{pre_human_test_and_treat}
\title{pre_human_test_and_treat}
\usage{
pre_human_test_and_treat(age_target, tail_states)
}
\arguments{
\item{age_target}{numeric of specific age groups to target/ select}

\item{tail_states}{output from Inter_run_processing}
}
\value{
tail states targeted for age-structured human T&T intervention
}
\description{
Takes processed tail states (following burn-in and a run) and selects specific age groups to implement human T&T moves
}
