% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{Pre_pig_MDA}
\alias{Pre_pig_MDA}
\title{Pre_pig_MDA}
\usage{
Pre_pig_MDA(age_target, Tail_states)
}
\arguments{
\item{age_target}{numeric of specific age groups to target/ select}

\item{Tail_states}{output from Inter_run_processing}
}
\value{
Tail states targetted for age-structured pig MDA intervention
}
\description{
Takes processed Tail states (following burn-in and a run) and selects specific age groups to implement pig MDA moves
}
