% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0.R
\name{r0}
\alias{r0}
\title{r0 wrapper}
\usage{
r0(params)
}
\arguments{
\item{params}{A list of model parameters: see \code{\link{r0_internal}} for list}
}
\description{
A wrapper function to facilitate the r0 function being called with a list of parameters
}
