% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0.R
\name{R0_internal}
\alias{R0_internal}
\title{R0 internal}
\usage{
R0_internal(
  delta,
  tau,
  phi,
  alpha,
  dH,
  dE,
  dP,
  RR_cysticercosis,
  theta,
  eta,
  chi,
  pih,
  pil,
  HPS,
  ...
)
}
\arguments{
\item{delta}{Egg production rate (per month)}

\item{tau}{Egg to pig transmission paramter}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}

\item{alpha}{Human recovery rate from Taenisasis (per month)}

\item{dH}{Human mortality rate (per month)}

\item{dE}{The egg mortlality/removal rate (per month)}

\item{dP}{Pig mortality rate (per month)}

\item{RR_cysticercosis}{Risk multiplier for Cysticercosis if human has Taeniasis}

\item{theta}{Egg to human transmission paramter}

\item{eta}{Human recovery rate from Cysticercosis (oer month)}

\item{chi}{rate of aquiring a pork meal (per month)}

\item{pih}{High intensity infected pig -> human contact rate (per month)}

\item{pil}{Low intensity infected pig -> human contact rate (per month)}

\item{HPS}{Human population size}
}
\description{
Calculate R0
}
