% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_wrappers.R
\name{Run_model}
\alias{Run_model}
\title{Run Cysticercosis model with interventions}
\usage{
Run_model(Params = NULL, Initial_states = NULL, Time, Intervention = NULL,
  Intervention_time = Time/2,
  Intervention_effect = Intervention_effect_size(), step = 1/30)
}
\arguments{
\item{Params}{List of model parameters}

\item{Initial_states}{List of intitial state values}

\item{Time}{The numebr of years to run the model for (from equilibrium). Default is at the halfway point.}

\item{Intervention}{A vector of interventions to include from: Husbandry, Sanitatio, Inspection, Pig_MDA, Pig_vaccine and Human_test_and_treat}

\item{Intervention_effect}{A list of intervention effect sizes, see \code{Intervention_effect_size} for details}

\item{step}{Time step (months)}
}
\description{
Runs the ODE Cysticercosis model
}
\examples{
M1<-Run_model(Time=75)
plot(M1$t/12, M1$Humans_Cysticercosis, t='l', ylim=c(0,1000), ylab='Humans with Cysticercosis', xlab='Time')
M2<-Run_model(Time=75, Intervention='Sanitation', Intervention_time=50)
head(M2)
lines(M2$t/12, M2$Humans_Cysticercosis, col='red')
M3<-Run_model(Time=75, Intervention=c('Human_test_and_treat', 'Pig_MDA'), Intervention_time=50)
lines(M3$t/12, M3$Humans_Cysticercosis, col='green')
legend('topright', c('Baseline','Sanitation','Human test & treat and Pig MDA'), lty=c(1,1,1), col=c('black','red', 'green'))

}

