% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up.R
\name{Set_up}
\alias{Set_up}
\title{Set up}
\usage{
Set_up(LEP = 1, delta = 960000, HPS = 10000, PPS = 2000, TPrev = 0.02,
  CPrev = 0.07, PCPrev = 0.2, AEL = 2, ATL = 2, ADI = 3, LEH = 54,
  phi = 0.8, chi = 0.5, RR_cysticercosis = 1, epsilon = 0.01,
  RR_infection = 2, RR_consumption = 0.75)
}
\arguments{
\item{LEP}{Pig life expectancy (years)}

\item{delta}{Egg production rate (per month)}

\item{HPS}{Human population size}

\item{PPS}{Pig population size}

\item{TPrev}{Taeniasis prevalence}

\item{CPrev}{Cysticercosis prevalence in humans}

\item{PCPrev}{Cysticercosis prevalence in pigs}

\item{AEL}{Average life expectancy of egg in the environment (weeks)}

\item{ATL}{Average life expectancy of T. solium adult worm in human (years)}

\item{ADI}{Average duration of cycticercosis infection in humans (years)}

\item{LEH}{Average life expectancy of a human (years)}

\item{phi}{Proportion of infected pigs with low-intensity cyst burden}

\item{chi}{Average number of pork meals (200g portion) per person per month}

\item{RR_cysticercosis}{Risk multiplier for Cysticercosis if human has Taeniasis}

\item{epsilon}{Pig rate of loss of naturally aquired immunity}

\item{RR_infection}{Increased RR of infection given consumption of high cyct burden pork, compared with low cyst burden}

\item{RR_consumption}{Decreased RR of consumption of high cyst burden pork, compared with low cyst burden pork}
}
\value{
Two lists of parameters and state variable values
}
\description{
Caculates internal parameters and equilbirum staring values for state variables
}

