% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_wrappers.R
\name{Single_run}
\alias{Single_run}
\title{Run Cysticercosis model ODE}
\usage{
Single_run(tt, params, states)
}
\arguments{
\item{tt}{vector of times}

\item{params}{list of parameters}

\item{states}{list of states}
}
\description{
Runs a single implementation of the ODE Cysticercosis model
}
