% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_up.R
\name{tau_equilibrium}
\alias{tau_equilibrium}
\title{Egg to pig transmission paramter}
\usage{
tau_equilibrium(dP, IP0, SP0, E0)
}
\arguments{
\item{dP}{Pig mortality rate (per month)}

\item{IP0}{Initial number of infected pigs}

\item{SP0}{Intitial number of susceptible pigs}

\item{E0}{The equilibrium number of eggs in the environment}
}
\value{
The equilibrium egg to pig transmission parameter
}
\description{
Calculate the transmission parameter (egg->pig) at equiliberium
}

