% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_processing.R
\name{Update_states}
\alias{Update_states}
\title{Update_states}
\usage{
Update_states(states_move, tail_states)
}
\arguments{
\item{states_move}{new age-specific states (updated following intervention moves)}

\item{tail_states}{output from Inter_run_processing (to update specific age-classes)}
}
\value{
overall Tail states list with new / updated specific age - classes following intervention moves
}
\description{
Identifies age targetted states and updates these specific states in the overall Tail states from the initial model run
}
